+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to the "amatl-vscode" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## 0.15.3
+
+- Add Amatl-Directives sidebar
+
+## 0.15.2
+
+- Construct of the README.md file
+
+## 0.15.1
 
 - Initial release